+++ aiobungie/crate/entity.py
@@ -35,6 +35,8 @@ __all__: tuple[str, ...] = (
     "GatingScope",
     "ValueUIStyle",
     "BaseEntity",
+    "ActivityEntity",
+    "PlaylistActivityEntity",
 )
 
 import abc
@@ -45,9 +47,12 @@ import attr
 from aiobungie.internal import enums
 
 if typing.TYPE_CHECKING:
+    import collections.abc as collections
+
     from aiobungie import traits
     from aiobungie import typedefs
     from aiobungie import undefined
+    from aiobungie.crate import activity
     from aiobungie.internal import assets
 
 
@@ -145,25 +150,25 @@ class BaseEntity(Entity):
     # implementation.
 
     net: traits.Netrunner = attr.field(repr=False)
-    # <<inherited docstring from Entity>>.
+    """A network state used for making external requests."""
 
     hash: int
-    # <<inherited docstring from Entity>>.
+    """Entity's hash."""
 
     index: int
-    # <<inherited docstring from Entity>>.
+    """The entity's index."""
 
     name: undefined.UndefinedOr[str]
-    # <<inherited docstring from Entity>>.
+    """Entity's name."""
 
     description: undefined.UndefinedOr[str]
-    # <<inherited docstring from Entity>>.
+    """Entity's description"""
 
     icon: assets.MaybeImage
-    # <<inherited docstring from Entity>>.
+    """Entity's icon."""
 
     has_icon: bool
-    # <<inherited docstring from Entity>>.
+    """A boolean that returns True if the entity has an icon."""
 
 
 @attr.define(kw_only=True, hash=False, weakref_slot=False)
@@ -282,3 +287,130 @@ class ObjectiveEntity(BaseEntity, Entity):
     perks: dict[str, int] = attr.field()
 
     stats: dict[str, int] = attr.field()
+
+
+@attr.define(kw_only=True, hash=True, weakref_slot=False)
+class ActivityEntity(BaseEntity, Entity):
+    """Represents a Bungie Activity definition and its entities.
+
+    This derives from `DestinyActivityDefinition` definition.
+    """
+
+    release_icon: assets.MaybeImage = attr.field(repr=False)
+    """The release icon of this activity if it has one."""
+
+    release_time: int = attr.field()
+    """The release time of this activity."""
+
+    unlock_hash: int = attr.field(repr=False)
+    """The completion unlock hash of this activity."""
+
+    light_level: int = attr.field()
+    """Activity's light level."""
+
+    place: typing.Union[
+        typedefs.IntAnd[enums.Place], typedefs.IntAnd[enums.Place]
+    ] = attr.field()
+    """The place of this activity."""
+
+    type_hash: int = attr.field(repr=False)
+    """The activity's type hash. This bounds to activity types such as Strikes, Crucible, Raids, etc."""
+
+    tier: typedefs.IntAnd[activity.Diffculity] = attr.field()
+    """Activity's diffculity tier."""
+
+    image: assets.MaybeImage = attr.field(repr=False)
+    """Activity's pgcr image."""
+
+    rewards: typing.Optional[collections.Sequence[activity.Rewards]] = attr.field()
+    """A sequence of this activity's rewards. Returns `None` if not found."""
+
+    modifiers: typing.Optional[collections.Sequence[int]] = attr.field(repr=False)
+    """A sequence of the activity's modifier hashes. Returns `None` if not found."""
+
+    challenges: typing.Optional[collections.Sequence[activity.Challenges]] = attr.field(
+        repr=False
+    )
+    """A sequence of the activity's challenges. Returns `None` if not found."""
+
+    is_playlist: bool = attr.field()
+    """Whether this activity is present in a playlist or not."""
+
+    unlock_strings: typing.Optional[collections.Sequence[str]] = attr.field(repr=False)
+    """An optional status string that could be conditionally displayed about an activity"""
+
+    inherits_free_room: bool = attr.field(repr=False)
+    """"""
+
+    playlist_activities: typing.Optional[
+        collections.Sequence[PlaylistActivityEntity]
+    ] = attr.field(repr=False)
+    """Represents all of the possible activities that could be played in the Playlist,
+    along with information that we can use to determine if they are active at the present time.
+    """
+
+    matchmaking: activity.Matchmaking = attr.field(repr=False)
+    """Information about matchmaking for this activity."""
+
+    guided_game: typing.Optional[activity.GuidedGame] = attr.field(repr=False)
+    """Information about activity's guided game mode, If exists otherwise `None`."""
+
+    mode: typing.Optional[typedefs.IntAnd[enums.GameMode]] = attr.field(repr=False)
+    """If this activity had an activity mode directly defined on it, this will be the hash of that mode."""
+
+    mode_hash: typing.Optional[int] = attr.field(repr=False)
+    """If the activity had an activity mode directly defined on it, this will be the enum value of that mode."""
+
+    mode_hashes: collections.Sequence[typedefs.IntAnd[enums.GameMode]] = attr.field(
+        repr=False
+    )
+    """The hash identifiers for Activity Modes relevant to this entry."""
+
+    mode_types: collections.Sequence[typedefs.IntAnd[enums.GameMode]] = attr.field(
+        repr=False
+    )
+    """A sequence of the activity gamemode types."""
+
+    loadouts: collections.Sequence[int] = attr.field(repr=False)
+    """The set of all possible loadout requirements that could be active for this activity.
+
+    Only one will be active at any given time. and you can discover which one through
+    activity-associated data such as Milestones that have activity info on them.
+    """
+
+    is_pvp: bool = attr.field(repr=False)
+    """Whether the activity is PvP or not."""
+
+    phase_hashes: collections.Sequence[int] = attr.field()
+    """The list of phases or points of entry into an activity,
+    along with information we can use to determine their gating and availability.
+    """
+
+    locations: collections.Collection[activity.Location] = attr.field(repr=False)
+    """A collection of location mappings affected by this activity."""
+
+
+@attr.define(kw_only=True, hash=True, weakref_slot=False)
+class PlaylistActivityEntity:
+    """Represents an activity playlists definition/entity.
+
+    Derives `DestinyActivityPlaylistItemDefinition`
+    """
+
+    hash: int = attr.field()
+    """The hash identifier of the Activity that can be played."""
+
+    mode_hash: typing.Optional[int] = attr.field()
+    """If this activity had an activity mode directly defined on it, this will be the hash of that mode."""
+
+    mode: typing.Optional[typedefs.IntAnd[enums.GameMode]] = attr.field()
+    """If the activity had an activity mode directly defined on it, this will be the enum value of that mode."""
+
+    mode_hashes: collections.Sequence[int] = attr.field()
+    """The hash identifiers for Activity Modes relevant to this entry."""
+
+    mode_types: collections.Sequence[typedefs.IntAnd[enums.GameMode]] = attr.field()
+    """A sequence of the activity gamemode types."""
+
+    async def fetch_self(self) -> ActivityEntity:
+        """Fetch the definition of this activivy."""